Fix fb#942
Added following new variable to service.xccfg:
# 1 to multiple 004 fields are present, move all but the first to valid 014$a fields
Replace014 = 0

This allows Voyager users to export bibs and holdings data separately, which is the safest
way to preserve holdings -> bibs relationships.  However, Voyager creates multiple 004 fields
to do this, which is not valid MARC.  This Normalization Service option allows us to
convert the invalid MARC into valid MARC by keeping the first 004 and creating 014$a fields
for the rest.




git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/trunk@2947 72d1a826-0e72-11de-b5bf-434079a3519f