In mst-common: No longer delete payload (MARCXML) when deleting records. It's useful to see this information when debugging, etc.
In MarcAggSvc:
Log all of the Record of Source data when debug=true. Previously only in-memory information was being logged; now everything is logged (in-memory *and* from db)

Fixed remerge(). If a deleted record happens to belong to a merge set, MAS attempted to merge the deleted record, but failed (NPE). We now delete merge data for the deleted record *before* remerging the set, and also not including it in the new merge set.



git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/trunk@3474 72d1a826-0e72-11de-b5bf-434079a3519f