Bug fix:
When a holdings record is "held" (references a bib via 004 and/or 014 fields), then becomes "active" ("unheld") due to the fact
that the holdings record was edited to remove a dangling reference, the marctoxctransformation service was
incorrectly storing the holding -> bib reference. This error caused future updates to incorrectly re-apply the
"held" status.


git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/trunk@3553 72d1a826-0e72-11de-b5bf-434079a3519f