Prepare for 1.5.5 relase.
Also, revised comment for previous commit:

When MarcAggSvc creates a new marged (deduped) record, it assumes that an 008 both EXISTS and is NOT EMPTY -- as it uses that as its key as the where to inject the new dynamic fields. However if this is not the case, it would simply concatenate this data to the end of the record AFTER the closing record tag. This would result in a malformed marcxml record! This change fixes that.